# demo/internal/router/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /auth/v1/login:
        post:
            tags:
                - Auth
            description: 登录
            operationId: Auth_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.auth.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.auth.v1.LoginResponse'
            security:
                - {}
    /auth/v1/logout:
        post:
            tags:
                - Auth
            description: 登出
            operationId: Auth_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.auth.v1.LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /auth/v1/me:
        get:
            tags:
                - Base
            description: 后台获取已经登录的用户的数据
            operationId: Base_GetMe
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.common.v1.User'
    /auth/v1/refresh_token:
        post:
            tags:
                - Auth
            description: 刷新认证令牌
            operationId: Auth_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.auth.v1.RefreshTokenRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.auth.v1.LoginResponse'
    /auth/v1/register:
        post:
            tags:
                - Auth
            operationId: Auth_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.auth.v1.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /helloworld/:name:
        get:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_SayHello
            parameters:
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloReply'
components:
    schemas:
        api.auth.v1.LoginRequest:
            required:
                - username
                - password
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                client_id:
                    type: string
                    description: 应用程序的识别ID
                client_secret:
                    type: string
                    description: 应用程序的密码
            description: 用户后台登录 - 请求
        api.auth.v1.LoginResponse:
            required:
                - access_token
                - refresh_token
            type: object
            properties:
                access_token:
                    type: string
                    description: 访问令牌，必选项。授权服务器颁发的访问令牌字符串。
                refresh_token:
                    type: string
                    description: 更新令牌，用来获取下一次的访问令牌，可选项。如果访问令牌将过期，则返回刷新令牌很有用，应用程序可以使用该刷新令牌来获取另一个访问令牌。但是，通过隐式授予颁发的令牌不能颁发刷新令牌。
                token_type:
                    type: string
                    default: Bearer
                    description: 令牌的类型，该值大小写不敏感，必选项，可以是bearer类型或mac类型，通常只是字符串“Bearer”。
                expires_in:
                    type: string
                    description: 令牌有效时间，单位为秒。如果访问令牌过期，服务器应回复授予访问令牌的持续时间。如果省略该参数，必须其他方式设置过期时间。
            description: 用户后台登录 - 回应
        api.auth.v1.LogoutRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
            description: 用户后台登出 - 请求
        api.auth.v1.RefreshTokenRequest:
            required:
                - refresh_token
            type: object
            properties:
                refresh_token:
                    type: string
                    description: 更新令牌，用来获取下一次的访问令牌，可选项。如果访问令牌将过期，则返回刷新令牌很有用，应用程序可以使用该刷新令牌来获取另一个访问令牌。但是，通过隐式授予颁发的令牌不能颁发刷新令牌。
                client_id:
                    type: string
                    description: 应用程序的识别ID
                client_secret:
                    type: string
                    description: 应用程序的密码
            description: 用户刷新令牌 - 请求
        api.auth.v1.RegisterRequest:
            required:
                - username
                - password
                - email
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                email:
                    type: string
        api.common.v1.User:
            type: object
            properties:
                id:
                    type: string
                    description: 用户ID
                roleId:
                    type: string
                    description: 角色ID
                createBy:
                    type: string
                    description: 创建者ID
                updateBy:
                    type: string
                    description: 更新者ID
                username:
                    type: string
                    description: 登录名
                password:
                    type: string
                    description: 密码
                nickname:
                    type: string
                    description: 昵称
                avatar:
                    type: string
                    description: 头像
                email:
                    type: string
                    description: 邮箱
                mobile:
                    type: string
                    description: 手机号
                gender:
                    type: integer
                    description: 性别
                    format: enum
                remark:
                    type: string
                    description: 备注名
                lastLoginTime:
                    type: string
                    description: 最后登录时间
                status:
                    enum:
                        - ON
                        - OFF
                    type: integer
                    default: ON
                    description: 用户状态
                    format: enum
                authority:
                    type: integer
                    default: CUSTOMER_USER
                    description: 权限
                    format: enum
                roles:
                    type: array
                    items:
                        type: string
                    description: 角色码
                createTime:
                    type: string
                    description: 创建时间
                updateTime:
                    type: string
                    description: 更新时间
            description: 用户
        helloworld.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
tags:
    - name: Auth
    - name: Base
    - name: Greeter
      description: The greeting service definition.